Fix concurrent modify bug
